server/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
cryptography==41.0.3
sqlalchemy==1.4.41
databases==0.6.2
asyncpg==0.28.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
